Fixes issue with stale objects when computing tax
If a user goes to the payment step with calculated tax on their order and then restarts the checkout, the old line item adjustment, containing the original tax charge, remains in memory. This results in the incorrect amount of tax getting applied to a completed order (leaving the order in a credit owed state).